Split FullDebug zlib/png into /MD and /MDd trees; quiet GRX/ZRX LNK4075.
Host-debug FullDebug (fulldebug.<family>.props) uses /MDd modules and *_mdd_*
static libs; default FullDebug stays /MD with *_md_*. Prepend host debug lib
dirs from props, apply /IGNORE:4099 on all modules, and force /INCREMENTAL:NO
for GRX/ZRX where SDK /GL forces LTCG. Also fix a few Studio warning cleanups
(ChmLib UTF-8, cast widths).

+++ Studio/ColumnsPage.cpp
@@ -189,7 +189,7 @@ BOOL CColumnsPage::OnInitDialog()
 			m_ColData[i].m_Width,
 			NULL,
 			0,
-			i,
+			static_cast<int>( i ),
 			m_ColData[i].m_Image,
 		};
 		if( m_ColData[i].m_Image >= 0 )
